refactor(rust): moved some helper code from examples to the ockam_identity crate
and reused a credential client hat already existed in ockam_api
