+++ examples/rust/get_started/src/lib.rs
@@ -7,14 +7,10 @@ mod hop;
 pub use forwarder::*;
 pub use hop::*;
 
-mod credentials;
-mod identity;
 mod logger;
 mod project;
 mod token;
 
-pub use credentials::*;
-pub use identity::*;
 pub use logger::*;
 pub use project::*;
 pub use token::*;
